pyproject.toml
[tool.poetry]
name = "toktik-backend"
version = "0.1.0"
description = ""
authors = ["mythicalbadger <jussy12.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.103.2"}
redis = "^5.0.1"
python-dotenv = "^1.0.0"
boto3 = "^1.28.64"
cryptography = "^41.0.4"
passlib = "^1.7.4"
sqlalchemy = "^2.0.22"
pyjwt = "^2.8.0"
psycopg2 = "^2.9.9"
python-socketio = "^5.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
